# src/main/java/com/erikj/pricing/infrastructure/adapter/in/rest/dto/PriceResponse.java
package com.erikj.pricing.infrastructure.adapter.in.rest.dto;

import com.erikj.pricing.domain.model.Price;

import java.math.BigDecimal;
import java.time.LocalDateTime;

public record PriceResponse(
        Long productId,
        Long brandId,
        Long priceList,
        LocalDateTime startDate,
        LocalDateTime endDate,
        BigDecimal price
) {

    public static PriceResponse fromDomain(Price price) {
        return new PriceResponse(
                price.productId(),
                price.brandId(),
                price.priceList(),
                price.startDate(),
                price.endDate(),
                price.price()
        );
    }
}